Feat: #108 SubscribeRepository 인터페이스 패턴 리팩토링
- SubscribeRepository를 인터페이스로 분리
- SubscribeRepositoryImpl 구현체 생성
- 구독 정보 조회 기능 인터페이스 적용
